+++ lib/models/user.js
@@ -17,7 +17,8 @@ const schema = new Schema({
 schema.static('exists', function (query) {
   return this.find(query)
     .count()
-    .then(count => (count));
+    // should return a boolean
+    .then(count => count > 0);
 });
 
 schema.method('generateHash', function (password) {
