+++ lib/routes/me.js
@@ -6,6 +6,8 @@ router
   .get('/', (req, res, next) => {
     User.findById(req.user.id)
       .lean()
+      // ACK! don't return the user's password hash
+      .select('-password')
       .then(user => res.send(user))
       .catch(next);
   });
