+++ lib/routes/posts.js
@@ -27,7 +27,13 @@ router
 
   .post('/:id/likes', (req, res, next) => {
     const id = req.params.id;
-    Post.findByIdAndUpdate(id, { $addToSet:  { likes: { $each: [req.user.id] } } }, { new: true })
+    Post.findByIdAndUpdate(id, { 
+      $addToSet:  { 
+        likes: { 
+          $each: [req.user.id] 
+        } 
+      }
+    }, { new: true })
       .then(post => res.send({likes: post.likes}))
       .catch(next);
   })
